Tighten integer types
Makes type more precise where they have to be unsigned anyway, are naturally limited to very small numbers, or where they're always used as collection indices.

+++ src/api/cover_calibrator.rs
@@ -81,8 +81,8 @@ pub trait CoverCalibrator: Device + Send + Sync {
     ///
     /// Only one interface version is current at a moment in time and all new devices should be built to the latest interface version. Applications can choose which device interface versions they support and it is in their interest to support  previous versions as well as the current version to ensure thay can use the largest number of devices.
     #[http("interfaceversion", method = Get)]
-    async fn interface_version(&self) -> ASCOMResult<i32> {
-        Ok(2_i32)
+    async fn interface_version(&self) -> ASCOMResult<u16> {
+        Ok(2)
     }
 }
