+++ src/api/dome.rs
@@ -161,8 +161,8 @@ pub trait Dome: Device + Send + Sync {
     ///
     /// Only one interface version is current at a moment in time and all new devices should be built to the latest interface version. Applications can choose which device interface versions they support and it is in their interest to support  previous versions as well as the current version to ensure thay can use the largest number of devices.
     #[http("interfaceversion", method = Get)]
-    async fn interface_version(&self) -> ASCOMResult<i32> {
-        Ok(3_i32)
+    async fn interface_version(&self) -> ASCOMResult<u16> {
+        Ok(3)
     }
 }
